+++ packages/@lou.codes/test/tests/default.test.ts
@@ -1,0 +1,9 @@
+import type { Test } from "../src/types/Test.js";
+
+// eslint-disable-next-line import/no-default-export
+export default {
+	given: "a default test",
+	must: "work",
+	received: () => true,
+	wanted: () => true,
+} satisfies Test;