+++ packages/@lou.codes/test/tests/multiple.test.ts
@@ -1,0 +1,10 @@
+import type { Tests } from "../src/types/Tests.js";
+
+export const multiple = [
+	{
+		given: "a test array",
+		must: "work",
+		received: () => true,
+		wanted: () => true,
+	},
+] satisfies Tests;