+++ packages/@lou.codes/test/tests/testImport/asyncIterator.ts
@@ -1,0 +1,3 @@
+export const asyncIterator = async function* () {
+	yield await Promise.resolve("test" as const);
+};