+++ packages/@lou.codes/test/tests/testImport/iterator.ts
@@ -1,0 +1,3 @@
+export const iterator = function* () {
+	yield "test";
+};