+++ packages/@lou.codes/create-package/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## 1.0.13
+
+### Patch Changes
+
+-   ⬆️ upgrade `@lou.codes/types` to `v1.1.4`
+-   ⬆️ upgrade `@lou.codes/ansi` to `v1.0.11`
+-   ⬆️ upgrade `@lou.codes/prompts` to `v1.0.11`
+
 ## 1.0.12
 
 ### Patch Changes