+++ packages/@lou.codes/predicates/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+## 1.0.11
+
+### Patch Changes
+
+-   ⬆️ upgrade `@lou.codes/types` to `v1.1.4`
+-   ⬆️ upgrade `@lou.codes/parsers` to `v1.0.11`
+
 ## 1.0.10
 
 ### Patch Changes